restructure examples testing
