+++ examples/autoignition.py
