--- examples/combozzle-mpi-lazy.py
@@ -1,1 +1,0 @@
-combozzle-mpi.py
