+++ examples/combozzle.py
