+++ examples/doublemach.py
