--- examples/doublemach_physical_av-mpi-lazy.py
@@ -1,1 +1,0 @@
-doublemach_physical_av-mpi.py
