+++ examples/hotplate.py
