+++ examples/mixture.py
