--- examples/multiple-volumes-mpi-lazy.py
@@ -1,1 +1,0 @@
-multiple-volumes-mpi.py
