--- examples/nsmix-mpi-lazy.py
@@ -1,1 +1,0 @@
-nsmix-mpi.py
