+++ examples/nsmix.py
