--- examples/poiseuille-local_dt-mpi-lazy.py
@@ -1,1 +1,0 @@
-poiseuille-local_dt-mpi.py
