+++ examples/poiseuille-local_dt.py
