+++ examples/scalar-lump.py
