--- examples/test_lazy_accuracy.sh
@@ -1,11 +1,0 @@
-#!/bin/bash
-
-set -x
-set -e
-set -o pipefail
-
-test_list="vortex-mpi.py pulse-mpi.py"
-for file in ${test_list}
-do
-    . ./compare_lazy_solution.sh ${file} 1e-13
-done
