--- examples/thermally-coupled-mpi-lazy.py
@@ -1,1 +1,0 @@
-thermally-coupled-mpi.py
