+++ examples/thermally-coupled.py
