+++ examples/vortex.py
